pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "jex jang",email = "jexjang92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "dotenv (>=0.9.9)",
    "livekit-agents[cartesia,deepgram,openai,silero,elevenlabs,turn-detector] (>=1.0.22)",
    "mem0ai (>=0.1.108)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "agent.main:main"